+++ src/componentTest/resources/expectedTemplateContent/add-permissions/rest-client/src/main/java/com/blackbaud/service/permissions/ServicePermissions.java
@@ -7,9 +7,7 @@ import java.util.Map;
 public class ServicePermissions extends AbstractPermissions {
     @Override
     protected void initPermissions(Map<String, Long> permissions) {
-        // TODO: Change these for your service as needed
-        permissions.put("service.view", 1L);
-        permissions.put("service.edit", 2L);
-        permissions.put("service.delete", 3L);
+        // TODO: Add permissions for your service as needed
+        // permissions.put("service.view", 1L);
     }
 }
