+++ src/componentTest/resources/expectedTemplateContent/add-permissions/src/main/java/com/blackbaud/service/core/CoreConfig.java
@@ -1,6 +1,5 @@
 package com.blackbaud.service.core;
 
-import com.blackbaud.security.GlobalMethodSecurityConfig;
 import com.blackbaud.feign.JacksonFeignBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
@@ -9,10 +8,6 @@ import org.springframework.context.annotation.Import;
 
 @Configuration
 @ComponentScan("com.blackbaud.service.core")
-@Import({
-        GlobalMethodSecurityConfig.class,
-        PermissionsConfig.class
-})
 public class CoreConfig {
 
     @Bean
