+++ src/main/resources/templates/springboot/rest/permissions/service-permissions.java.tmpl
@@ -7,9 +7,7 @@ import java.util.Map;
 public class ${serviceName}Permissions extends AbstractPermissions {
     @Override
     protected void initPermissions(Map<String, Long> permissions) {
-        // TODO: Change these for your service as needed
-        permissions.put("${servicePermissionPrefix}.view", 1L);
-        permissions.put("${servicePermissionPrefix}.edit", 2L);
-        permissions.put("${servicePermissionPrefix}.delete", 3L);
+        // TODO: Add permissions for your service as needed
+        // permissions.put("${servicePermissionPrefix}.view", 1L);
     }
 }
